+++ objc2-encode/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+
+## 2.0.0-pre.1 - 2022-07-19
+
 ### Added
 * Added `Encoding::Atomic`.
 * Implement `Encode` and `RefEncode` for `std::sync::atomic` types.